WIP add entry point
It is very dumb. This would save us from having to publish 4 or 5 binaries.

Instead of `iroh-gateway ...`, you do `iroh gateway ...`.

```
Usage: iroh [COMMAND]

Commands:
  gateway  IPFS gateway
  p2p      Implementation of the p2p part of iroh
  store    Implementation of the storage part of iroh
  cli      A next generation IPFS implementation: https://iroh.computer
  help     Print this message or the help of the given subcommand(s)

Options:
  -h, --help     Print help information
  -V, --version  Print version information
```
