+++ Cargo.toml
@@ -14,6 +14,7 @@ members = [
   "iroh-store",
   "iroh-share",
   "iroh-util",
+  "iroh-entry-point",
   "stores/*",
   "examples",
   "xtask"
