+++ iroh-gateway/src/main.rs
@@ -1,18 +1,11 @@
-use std::sync::Arc;
-
-use anyhow::{anyhow, Result};
+use anyhow::Result;
 use clap::Parser;
 use iroh_gateway::{
-    bad_bits::{self, BadBits},
     cli::Args,
     config::{Config, CONFIG_FILE_NAME, ENV_PREFIX},
-    core::Core,
-    metrics,
+    metrics, run,
 };
-use iroh_rpc_client::Client as RpcClient;
 use iroh_util::{iroh_config_path, make_config};
-use tokio::sync::RwLock;
-use tracing::{debug, error};
 
 #[tokio::main(flavor = "multi_thread")]
 async fn main() -> Result<()> {
@@ -33,51 +26,5 @@ async fn main() -> Result<()> {
     .unwrap();
     config.metrics = metrics::metrics_config_with_compile_time_info(config.metrics);
     println!("{:#?}", config);
-
-    let metrics_config = config.metrics.clone();
-    let bad_bits = match config.denylist {
-        true => Arc::new(Some(RwLock::new(BadBits::new()))),
-        false => Arc::new(None),
-    };
-    let rpc_addr = config
-        .server_rpc_addr()?
-        .ok_or_else(|| anyhow!("missing gateway rpc addr"))?;
-    let content_loader = RpcClient::new(config.rpc_client.clone()).await?;
-    let handler = Core::new(
-        Arc::new(config),
-        rpc_addr,
-        Arc::clone(&bad_bits),
-        content_loader,
-    )
-    .await?;
-
-    let bad_bits_handle = bad_bits::spawn_bad_bits_updater(Arc::clone(&bad_bits));
-
-    let metrics_handle = iroh_metrics::MetricsHandle::new(metrics_config)
-        .await
-        .expect("failed to initialize metrics");
-
-    #[cfg(unix)]
-    {
-        match iroh_util::increase_fd_limit() {
-            Ok(soft) => debug!("NOFILE limit: soft = {}", soft),
-            Err(err) => error!("Error increasing NOFILE limit: {}", err),
-        }
-    }
-
-    let server = handler.server();
-    println!("listening on {}", server.local_addr());
-    let core_task = tokio::spawn(async move {
-        server.await.unwrap();
-    });
-
-    iroh_util::block_until_sigint().await;
-    core_task.abort();
-
-    metrics_handle.shutdown();
-    if let Some(handle) = bad_bits_handle {
-        handle.abort();
-    }
-
-    Ok(())
+    run(config).await
 }
