+++ iroh-p2p/src/lib.rs
@@ -5,7 +5,9 @@ mod keys;
 pub mod metrics;
 mod node;
 pub mod rpc;
+mod run;
 mod swarm;
+pub use run::run;
 
 pub use self::config::*;
 pub use self::keys::{DiskStorage, Keychain, MemoryStorage};
