+++ iroh-store/src/lib.rs
@@ -3,7 +3,9 @@ pub mod cli;
 pub mod config;
 pub mod metrics;
 pub mod rpc;
+mod run;
 mod store;
+pub use run::run;
 
 pub use crate::config::Config;
 pub use crate::store::Store;
